requirements.txt
requests
gensim
python-Levenshtein
titlecase
nltk
flask
twitter
pyyaml